+++ CHANGELOG.md
@@ -17,6 +17,7 @@ support.
 
 ### Changed
 - Synchronize source code with standard library of Rust version 1.87.0.
+- When `extra` is enabled, `PeekMut::refresh` is potentially faster.
 
 ## [0.3.7] &ndash; 2024-10-18
 ### Added